+++ generator/opcodes/Decoder.cs
@@ -17,7 +17,7 @@ namespace emulator
         public CPU(Func<byte> read) : this(() => 0, x => { }, x => { }, new MMU(read))
         { }
 
-        public CPU(Func<ushort> getPC, Action<ushort> setPC, Action<int> TickClock, MMU memory)
+        public CPU(Func<ushort> getPC, Action<ushort> setPC, Action<long> TickClock, MMU memory)
         {
             StdOps = MakeTable();
             CbOps = MakeTableCb();
